% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_to_in.R
\name{mm_to_in}
\alias{mm_to_in}
\title{mm to inches converter}
\usage{
mm_to_in(x)
}
\arguments{
\item{x}{measurement in mm provided. (mm)}
}
\value{
unit conversion of mm into inches.
}
\description{
The purpose of this function is to convert mm into inches, by the conversion factor of 1/254.
}
\examples{
mm_to_in(200) # should be "0.7874016"
mm_to_in(165.2) # should be "0.6503937"
mm_to_in(-1933.3) # should be "-7.611417"
}
